requirements.txt
fastapi==0.116.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
boto3==1.40.11
python-jose==3.5.0
passlib==1.7.4
pydantic==2.11.7
gspread==6.0.0
google-auth==2.39.0
python-multipart==0.0.6
email-validator==2.2.0
requests==2.31.0
bcrypt==3.2.0
pyjwt==2.10.1
email-validator==2.2.0
httpx>=0.24.0
pymongo==4.8.0
motor==3.5.1